Added function getProjectStatusPl.
Because AHK is stupid and wouldn't accept a command line argument to
stop popping dialogs asking to create a missing script file, the
existence of the script file needs to be checked in PHP first.
While I was at it, I refactored the whole section to something readable.
I ignored that RingDat_en's script file is named "Test starten -
RingDat_EN.ahk", ditto for 'de', because file names are case-insensitive
in Windows.