Made a broken solution slightly less broken.
I was viewing differences in a 'details' page, when suddenly the page
said 'Currently no target state' after I reloaded the tab; The status
was wrong, as the target state was still in the images folder.
The fix introduced in b18c9426 breaks <image src> attributes in
details.tpl and presumably addRtf()'s link.
I restricted the fix to actual long file names to honor the author's
good intentions.